+++ PageManagements/Assets/PageManagements/Scripts/Runtime/IPage.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Threading;
+using Cysharp.Threading.Tasks;
+
+namespace PageManagements
+{
+    public interface IPage : IDisposable
+    {
+        UniTask Show(CancellationToken cancellationToken);
+        UniTask Hide(CancellationToken cancellationToken);
+    }
+}