+++ PageManagements/Assets/PageManagements/Scripts/Runtime/PageArgument.cs
@@ -1,8 +1,0 @@
-namespace PageManagements
-{
-    public class PageArgument
-    {
-        public PageBuilder PageBuilder { get; set; }
-        public PageManager PageManager { get; set; }
-    }
-}