+++ PageManagements/Assets/PageManagements/Scripts/Runtime/PageManager.cs
@@ -1,58 +1,81 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
+using Cysharp.Threading.Tasks;
+using UnityEngine;
 
 namespace PageManagements
 {
     public class PageManager
     {
-        private List<PageBase> _pages = new List<PageBase>();
+        private readonly PageBuilder _pageBuilder;
+        private readonly List<IPage> _pages = new List<IPage>();
 
         public event Action PageChanged;
 
-        public void Push(PageBase page)
+        public PageManager(PageBuilder pageBuilder)
         {
-            var oldPage = _pages.LastOrDefault();
-
-            // TODO: Add transition
-            if (oldPage == null)
-            {
-                page.Show();
-            }
-            else
-            {
-                oldPage.Hide();
-                page.Show();
-            }
-
-            _pages.Add(page);
-            PageChanged?.Invoke();
+            _pageBuilder = pageBuilder;
         }
 
-        public void Pop()
+        public async UniTask<PageHandle<T>> Create<T>(CancellationToken cancellationToken) where T : IPage
         {
-            if (_pages.Count == 0)
+            var page = _pageBuilder.Build<T>();
+            await Push(page, cancellationToken);
+            return new PageHandle<T>(page, this);
+        }
+
+        private async UniTask Push(IPage page, CancellationToken cancellationToken)
+        {
+            var oldPage = _pages.LastOrDefault();
+            _pages.Add(page);
+            PageChanged?.Invoke();
+            // Switch page animation
+            if (oldPage != null)
             {
+                await oldPage.Hide(cancellationToken);
+            }
+            await page.Show(cancellationToken);
+        }
+
+        public UniTask Remove(IPage page, CancellationToken cancellationToken)
+        {
+            var index = _pages.IndexOf(page);
+            if (index == -1)
+            {
+                Debug.LogError($"Page not found. {page}");
+                return UniTask.CompletedTask;
+            }
+            return Remove(index, cancellationToken);
+        }
+
+        public async UniTask Remove(int index, CancellationToken cancellationToken)
+        {
+            if (index < 0 || index >= _pages.Count)
+            {
+                Debug.LogError($"Invalid index. pages:{_pages.Count}, index:{index}");
                 return;
             }
 
-            var lastPage = _pages.Last();
-            var prevPage = _pages.Count > 1 ? _pages[_pages.Count - 2] : null;
-
-            // TODO: Add transition
-            if (prevPage == null)
-            {
-                lastPage.Hide();
-            }
-            else
-            {
-                lastPage.Hide();
-                prevPage.Show();
-            }
-
-            _pages.Remove(lastPage);
-            lastPage.Dispose();
+            var isLastPage = index == _pages.Count - 1;
+            var page = _pages[index];
+            _pages.Remove(page);
             PageChanged?.Invoke();
+
+            // Switch page animation
+            await page.Hide(cancellationToken);
+            page.Dispose();
+            if (isLastPage && _pages.Count > 0)
+            {
+                var prevPage = _pages.Last();
+                await prevPage.Show(cancellationToken);
+            }
+        }
+
+        public UniTask Pop(CancellationToken cancellationToken)
+        {
+            return Remove(_pages.Count - 1, cancellationToken);
         }
 
         public int GetPageCount()
@@ -60,7 +83,7 @@ namespace PageManagements
             return _pages.Count;
         }
 
-        public bool HasPage<T>() where T : PageBase
+        public bool HasPage<T>() where T : IPage
         {
             var pageCount = _pages.Count;
             for (var i = 0; i < pageCount; i++)