+++ PageManagements/Assets/PageManagements/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.1.0] - 2024-07-22
+### Changed
+- Disruptive changes:
+- Changed the PageBase class to the IPage interface.
+- Changed to a format that calls the function of the Page instance instead of passing PageArgument.
+- Changed so that PageManager generates pages using PageBuilder.
+- Changed so that PageManager returns PageHandle.
+
 ## [1.0.1] - 2024-07-11
 ### Added
 - Added samples.